Refactored project name, initial code setup
