+++ pom.xml
@@ -9,10 +9,10 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>org.example</groupId>
-    <artifactId>DungeonsAndDebugerss</artifactId>
+    <artifactId>HibernateHaunters</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>DungeonsAndDebugerss</name>
-    <description>DungeonsAndDebugerss</description>
+    <name>HibernateHaunters</name>
+    <description>HibernateHaunters</description>
     <properties>
         <java.version>21</java.version>
     </properties>
@@ -39,6 +39,14 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-rest</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
     </dependencies>
