--- src/main/java/org/example/dungeonsanddebugerss/respositories/CountrylanguageEntityRepository.java
@@ -1,8 +1,0 @@
-package org.example.dungeonsanddebugerss.respositories;
-
-import org.example.dungeonsanddebugerss.entities.CountrylanguageEntity;
-import org.example.dungeonsanddebugerss.entities.CountrylanguageEntityId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CountrylanguageEntityRepository extends JpaRepository<CountrylanguageEntity, CountrylanguageEntityId> {
-}
