+++ src/main/java/org/example/hibernatehaunters/controller/CountryController.java
@@ -1,0 +1,4 @@
+package org.example.hibernatehaunters.controller;
+
+public class CountryController {
+}
