+++ src/main/java/org/example/hibernatehaunters/controller/CountrylanguageController.java
@@ -1,0 +1,5 @@
+package org.example.hibernatehaunters.controller;
+
+
+public class CountrylanguageController {
+}
