+++ src/main/java/org/example/hibernatehaunters/models/respositories/CityEntityRepository.java
@@ -1,6 +1,6 @@
-package org.example.dungeonsanddebugerss.respositories;
+package org.example.hibernatehaunters.models.respositories;
 
-import org.example.dungeonsanddebugerss.entities.CityEntity;
+import org.example.hibernatehaunters.models.entities.CityEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CityEntityRepository extends JpaRepository<CityEntity, Integer> {
