+++ src/main/java/org/example/hibernatehaunters/models/respositories/CountryEntityRepository.java
@@ -1,6 +1,6 @@
-package org.example.dungeonsanddebugerss.respositories;
+package org.example.hibernatehaunters.models.respositories;
 
-import org.example.dungeonsanddebugerss.entities.CountryEntity;
+import org.example.hibernatehaunters.models.entities.CountryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CountryEntityRepository extends JpaRepository<CountryEntity, String> {
