+++ src/main/java/org/example/hibernatehaunters/service/CityService.java
@@ -1,8 +1,7 @@
-package org.example.dungeonsanddebugerss.service;
+package org.example.hibernatehaunters.service;
 
-import org.example.dungeonsanddebugerss.entities.CityEntity;
-import org.example.dungeonsanddebugerss.respositories.CityEntityRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.example.hibernatehaunters.models.entities.CityEntity;
+import org.example.hibernatehaunters.models.respositories.CityEntityRepository;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
