+++ src/main/java/org/example/hibernatehaunters/service/CountryService.java
@@ -1,7 +1,7 @@
-package org.example.dungeonsanddebugerss.service;
+package org.example.hibernatehaunters.service;
 
-import org.example.dungeonsanddebugerss.entities.CountryEntity;
-import org.example.dungeonsanddebugerss.respositories.CountryEntityRepository;
+import org.example.hibernatehaunters.models.entities.CountryEntity;
+import org.example.hibernatehaunters.models.respositories.CountryEntityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
