--- src/test/java/org/example/dungeonsanddebugerss/service/CountrylanguageServiceTests.java
@@ -1,5 +1,0 @@
-package org.example.dungeonsanddebugerss.service;
-
-public class CountrylanguageServiceTests {
-
-}
