+++ src/test/java/org/example/hibernatehaunters/service/CityServiceTests.java
@@ -1,8 +1,8 @@
-package org.example.dungeonsanddebugerss.service;
+package org.example.hibernatehaunters.service;
 
-import org.example.dungeonsanddebugerss.entities.CityEntity;
-import org.example.dungeonsanddebugerss.entities.CountryEntity;
-import org.example.dungeonsanddebugerss.respositories.CityEntityRepository;
+import org.example.hibernatehaunters.models.entities.CityEntity;
+import org.example.hibernatehaunters.models.entities.CountryEntity;
+import org.example.hibernatehaunters.models.respositories.CityEntityRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
