+++ src/test/java/org/example/hibernatehaunters/service/CountryServiceTests.java
@@ -1,7 +1,7 @@
-package org.example.dungeonsanddebugerss.service;
+package org.example.hibernatehaunters.service;
 
-import org.example.dungeonsanddebugerss.entities.CountryEntity;
-import org.example.dungeonsanddebugerss.respositories.CountryEntityRepository;
+import org.example.hibernatehaunters.models.entities.CountryEntity;
+import org.example.hibernatehaunters.models.respositories.CountryEntityRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
