+++ src/test/java/org/example/hibernatehaunters/service/CountrylanguageServiceTests.java
@@ -1,0 +1,5 @@
+package org.example.hibernatehaunters.service;
+
+public class CountrylanguageServiceTests {
+
+}
